requirements.txt
python-libnmap>=0.7.0
netaddr>=0.8.0
argparse>=1.4.0
scapy>=2.4.5